refactor: step3 피드백, 구간 추가 인수테스트 Line 인수 테스트와 분리
